+++ CHANGELOG.md
@@ -18,7 +18,9 @@ configured prefix accordingly.
 ### Added
 
 - Added debug bot commands `-argstr` and `-argint`
+- Added bot commands `-repo` and `-issues`
 - Added command methods `cogs.debug.DebugCog.argstr()` and `cogs.debug.DebugCog.argint()`
+- Added command methods `cogs.general.GeneralCog.repo()` and `cogs.general.GeneralCog.issues()`
 
 ### Changed